# apps/api/gpt/openapi.yaml
openapi: 3.1.0
info:
  title: TimTracker Query API
  description: |
    API for querying TimTracker tables and summarizing health data. 
    Exposes both raw table access and analytic endpoints. 
    Use /api/schema to discover tables, columns, and endpoints. 
    For most use cases, prefer analytic endpoints (e.g., /api/sleep) over raw tables.
  version: '2.0'
servers:
  - url: https://timtracker-api.vercel.app
paths:
...
  /api/log/meal:
    post:
      summary: Log a meal
      operationId: logMeal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                date:
                  type: string
                  format: date
                health_score:
                  type: integer
                health_comment:
                  type: string
              required:
                - text
                - date
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Meal logged
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
  /api/log/meal/{id}:
    patch:
      summary: Update a previously logged meal
      operationId: updateMeal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                date:
                  type: string
                  format: date
                health_score:
                  type: integer
                health_comment:
                  type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Meal updated

    delete:
      summary: Delete a previously logged meal
      operationId: deleteMeal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Meal deleted
...
  /api/interactions:
    get:
      summary: List interactions joined with people names
      operationId: listInteractions
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - in: query
          name: start
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: end
          required: false
          schema:
            type: string
            format: date
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of interactions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    date:
                      type: string
                      format: date
                    person_name:
                      type: string
                    interaction_type:
                      type: string
                      enum: [IRL, Call, Text]
                    note:
                      type: string
  /api/log/interaction:
    post:
      summary: Log one or more social interactions
      operationId: logInteraction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                interactions:
                  type: array
                  items:
                    type: object
                    properties:
                      people_ids:
                        type: array
                        items:
                          type: integer
                      people_notes:
                        type: object
                        additionalProperties:
                          type: string
                      date:
                        type: string
                        format: date
                      type:
                        type: string
                        enum: [IRL, Call, Text]
                      note:
                        type: string
                    required:
                      - people_ids
                      - date
                      - type
              required:
                - interactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Interactions logged
          content:
            application/json:
              schema:
                type: object
                properties:
                  ids:
                    type: array
                    items:
                      type: integer
  /api/log/interaction/{id}:
    patch:
      summary: Update a logged social interaction
      operationId: updateInteraction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                people_ids:
                  type: array
                  items:
                    type: integer
                people_notes:
                  type: object
                  additionalProperties:
                    type: string
                date:
                  type: string
                  format: date
                type:
                  type: string
                  enum: [IRL, Call, Text]
                note:
                  type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Interaction updated

    delete:
      summary: Delete a social interaction
      operationId: deleteInteraction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Interaction deleted
  /api/daily-meal-scores:
    get:
      summary: Daily meal health scores
      operationId: getDailyMealScores
      parameters:
        - in: query
          name: start
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: end
          required: false
          schema:
            type: string
            format: date
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of daily scores
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                    health_score:
                      type: integer
                    health_comment:
                      type: string

    post:
      summary: Upsert a daily meal score
      operationId: upsertDailyMealScore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                health_score:
                  type: integer
                health_comment:
                  type: string
              required:
                - date
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Score saved
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token required to authenticate requests. 
        Use the GPT API key configured in the ChatGPT connector settings.
...